[FEAT] hideKeyboard 공용화 https://github.com/POLZZAK/POLZZAK-Android/issues/34
