+++ app/src/main/java/com/polzzak_android/presentation/common/util/FragmentUtil.kt
@@ -1,5 +1,7 @@
 package com.polzzak_android.presentation.common.util
 
+import android.content.Context
+import android.view.inputmethod.InputMethodManager
 import androidx.fragment.app.Fragment
 import com.polzzak_android.presentation.auth.login.sociallogin.SocialLoginManager
 import com.polzzak_android.presentation.common.MainActivity
@@ -10,4 +12,16 @@ fun Fragment.getAccessTokenOrNull(): String? = (activity as? MainActivity)?.getA
 
 fun Fragment.shotBackPressed() {
     (activity as? MainActivity)?.backPressed()
+}
+
+fun Fragment.hideKeyboard() {
+    activity?.run {
+        val inputManager = getSystemService(Context.INPUT_METHOD_SERVICE) as? InputMethodManager
+        currentFocus?.let { currentFocus ->
+            inputManager?.hideSoftInputFromWindow(
+                currentFocus.windowToken,
+                InputMethodManager.HIDE_NOT_ALWAYS
+            )
+        }
+    }
 }
